US-09 move methods for tags to its dataclasses
